+++ tabs/plot_pango.c
@@ -1,4 +1,5 @@
 
+#include <cglm/cglm.h>
 #include "plot_pango.h"
 
 static PangoFontDescription *plot_font;
@@ -14,7 +15,7 @@ static cairo_t* plot_create_cairo(void) {
   return cairo;
 }
 
-static cairo_t* plot_create_cairo_from_buff(ivec2_t size, cairo_format_t format,
+static cairo_t* plot_create_cairo_from_buff(int size[2], cairo_format_t format,
     cairo_surface_t **psurf, unsigned char **pbuff) {
   if (!pbuff || !psurf) return NULL;
   int stride = cairo_format_stride_for_width(format, size[0]);
@@ -28,7 +29,7 @@ static cairo_t* plot_create_cairo_from_buff(ivec2_t size, cairo_format_t format,
   return NULL;
 }
 
-static GLuint plot_create_texture(ivec2_t size, const void *pixels) {
+static GLuint plot_create_texture(int size[2], const void *pixels) {
   GLuint id = 0;
   glGenTextures(1, &id);
   if (id) {
@@ -64,7 +65,7 @@ void plot_pango_free(void) {
   if (plot_font) { pango_font_description_free(plot_font); plot_font = NULL; }
 }
 
-GLuint plot_pango_text(char ch, ivec2_t size) {
+GLuint plot_pango_text(char ch, int size[2]) {
   if (!plot_font) return 0;
   GLuint tid = 0;
   cairo_t *cairo = plot_create_cairo();