+++ ui/action.h
@@ -14,9 +14,11 @@ enum { ACCL_SA_LGND,
 ACCL_SA_MAX };
 
 #ifdef WITH_PLOT
+typedef struct kb_rot_lg { const t_ent_spn_aux *aux; int *pval, sign; gboolean rev; } t_kb_rot_lg;
+
 typedef struct kb_rotaux {
-  const t_ent_spn_aux *aux;
-  int ndx, *pval, inc, *scale;
+  t_kb_rot_lg global, local;
+  int label, *step;
 } t_kb_rotaux;
 extern t_kb_rotaux kb_rotaux[ACCL_SA_MAX];
 void on_rotation(GSimpleAction*, GVariant*, t_kb_rotaux*);